requirements.txt
altair==5.5.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
black==22.12.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
crispy-bootstrap5==0.7
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
data==0.4
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
drf-yasg==1.21.10
entrypoints==0.4
factory_boy==3.3.1
Faker==37.1.0
favicon==0.7.0
firebase-admin==6.8.0
fonttools==4.57.0
funcsigs==1.0.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gopay==2.0.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
htbuilder==0.9.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
latex==0.7.0
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
mypy_extensions==1.1.0
narwhals==1.37.1
numpy==2.2.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
plotly==6.0.1
prettytable==3.16.0
prometheus_client==0.21.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydeck==0.9.1
pyecharts==2.0.8
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pymdown-extensions==10.15
pyparsing==3.2.3
pypdf==5.4.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2022.7
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
reportlab==4.0.4
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
shutilwhich==1.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.3
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.45.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
stripe==5.5.0
svglib==1.5.1
sympy==1.14.0
tempdir==0.7.1
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.17.2
xhtml2pdf==0.2.15
